+++ src/components/Radio/Radio.tsx
@@ -111,10 +111,10 @@ export const Radio = chakra(
         </ChakraRadio>
         {footnote && showHelperInvalidText && (
           <HelperErrorText
-            additionalStyles={styles.helperErrorText}
             id={`${id}-helperText`}
             isInvalid={isInvalid}
             text={footnote}
+            __css={styles.helperErrorText}
           />
         )}
       </>