+++ src/components/Select/Select.tsx
@@ -14,8 +14,6 @@ import { IconNames, IconSizes } from "../Icons/IconTypes";
 import Label from "../Label/Label";
 import { SelectTypes, LabelPositions } from "./SelectTypes";
 export interface SelectProps {
-  /** Optionally pass in additional Chakra-based styles. */
-  additionalStyles?: { [key: string]: any };
   /** A class name for the `div` parent element. */
   className?: string;
   /** Optional string to populate the `HelperErrorText` for the standard state. */
@@ -69,7 +67,6 @@ export const Select = chakra(
   React.forwardRef<HTMLSelectElement, React.PropsWithChildren<SelectProps>>(
     (props: React.PropsWithChildren<SelectProps>, ref?) => {
       const {
-        additionalStyles = {},
         children,
         className,
         helperText,
@@ -131,11 +128,7 @@ export const Select = chakra(
       }, [labelPosition]);
 
       return (
-        <Box
-          className={className}
-          __css={{ ...styles, ...additionalStyles }}
-          {...rest}
-        >
+        <Box className={className} __css={styles} {...rest}>
           <Box __css={labelPosition === LabelPositions.Inline && styles.inline}>
             {showLabel && (
               <Box ref={labelRef}>